dark_halo_scope_2/requirements.txt
astropy==6.0.1
astropy-iers-data==0.2025.12.1.0.45.12
black==25.11.0
boto3>=1.34.0
botocore>=1.34.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.8
configparser==7.2.0
contourpy==1.3.0
corner==2.2.3
coverage==7.10.7
cycler==0.12.1
-e git+ssh://git@github.com/aditrivaishnavi/toy_glens_kinematics.git@c807f0fc86a5aff4733e342ada8f8d711a46dd10#egg=dark_halo_scope&subdirectory=dark_halo_scope
dill==0.4.0
exceptiongroup==1.3.1
filelock==3.19.1
fonttools==4.60.1
fsspec==2025.10.0
h5py==3.14.0
idna==3.11
ImageIO==2.37.2
importlib_resources==6.5.2
iniconfig==2.1.0
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.7
lazy_loader==0.4
lenstronomy==1.13.3
llvmlite==0.43.0
MarkupSafe==3.0.3
matplotlib==3.9.4
mpmath==1.3.0
multiprocess==0.70.18
mypy_extensions==1.1.0
networkx==3.2.1
numba==0.60.0
numpy==1.26.4
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.4.0
pluggy==1.6.0
pyerfa==2.0.1.5
Pygments==2.19.2
pyparsing==3.2.5
pytest==8.4.2
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
pytokens==0.3.0
pytz==2025.2
pyvo==1.8
pyxdg==0.28
PyYAML==6.0.3
requests==2.32.5
ruff==0.14.8
schwimmbad==0.4.2
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2024.8.30
tomli==2.3.0
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.0
zipp==3.23.0
